\name{dew_temperature_from_temperature_RH}
\alias{dew_temperature_from_temperature_RH}
\title{It calculates dew temperature from air temperature and relative humidity}
\usage{
dew_temperature_from_temperature_RH(T, RH, percent = TRUE, ...)
}
\arguments{
  \item{T}{temperature}

  \item{RH}{relative humidity}

  \item{percent}{logical value. If \code{TRUE} (default)
  the relative humidity is returned in per cent}

  \item{...}{arguments for
  \code{\link{temperature_from_saturated_water_vapor_pressure}}}
}
\description{
It calculates dew temperature from air temperature and
relative humidity
}

