\name{pine}
\docType{data}
\alias{pine}

\title{
Data set for RHUMIDITY: pine
}

\usage{
data(pine)
data(gen_2001_2002_00001)
}

\description{
It contains the following variables: 
\describe{

\item{\code{pine}}{\code{zoo} object containing hourly time-series of temparature[Celsius Degrees], relative humidity [Percent], precipitation [mm], leaf wetness (actually incomplete) at Baselga di Pine (Trentino, Italy)  }                 
%}
%%\item{\code{pine}}{\code{zoo} object containing hourly time-series of temparature[Celsius Degrees], relative humidity [Percent], precipitation [mm], leaf  TO DO Data frame containing \code{year},\code{month} , \code{day} and daily maximum temperature in 59 stations in Trentino region}                 
}
}
\details{
This data set is an example of hourly data to be used for the package eexamples. The time series of leaf wetness may be incorrect and will be updated. 
The user can easily use the package with his/her own data after replacing the values of such variables.
}
\format{
zoo oblects
}
\source{
Agro-Meteorological Office Technological Transfer CEnter, Fondazione Edmund Mach \url{http://meteo.fmach.it/meteo/}
This dataset is intended for research purposes only, being distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY.
}

%%\references{
%% ~put references to the literature/web site here ~
%%}
\keyword{dataset}
